Restructure list card into appropriate folder
