+++ movie_shape/lib/reusable/list/list_card.dart
